sanctum

+++ app/Http/Controllers/ProductController.php
@@ -14,7 +14,7 @@ class ProductController extends Controller
      */
     public function index()
     {
-        //read all products
+        // Read all products
         return Product::all();
     }
 
@@ -26,20 +26,28 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required|min:5',
-            'slug' => 'required',
-            'price' => 'required'
-            
-        ]);
+        // เช็คสิทธิ์ (role) ว่าเป็น admin (1) 
+        $user = auth()->user();
 
-        return Product::create($request->all());
+        if($user->tokenCan("1")){
+            $request->validate([
+                'name' => 'required|min:5',
+                'slug' => 'required',
+                'price' => 'required'
+            ]);
+
+            return Product::create($request->all());
+        }else{
+            return [
+                'status' => 'Permission denied to create'
+            ];
+        }
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Product  $product
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
@@ -56,19 +64,40 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $product = Product::find($id);
-        $product->update($request->all());
-        return $product;
+        // เช็คสิทธิ์ (role) ว่าเป็น admin (1) 
+        $user = auth()->user();
+
+        if($user->tokenCan("1")){
+            $product = Product::find($id);
+            $product->update($request->all());
+
+            return $product;
+        }else{
+            return [
+                'status' => 'Permission denied to create'
+            ];
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Product  $product
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        return Product::destroy($id);
+        
+        // เช็คสิทธิ์ (role) ว่าเป็น admin (1) 
+        $user = auth()->user();
+
+        if($user->tokenCan("1")){
+            return Product::destroy($id);
+        }else{
+            return [
+                'status' => 'Permission denied to create'
+            ];
+        }
+
     }
 }
